Use configure on demand
