+++ settings.gradle.kts
@@ -13,9 +13,9 @@ setupSubproject("$forkNameLowercase-server") {
     projectDir = File("$forkName-Server")
     buildFileName = "../subprojects/server.gradle.kts"
 }
+setupSubproject("Yatoclip") { }
 
 inline fun setupSubproject(name: String, block: ProjectDescriptor.() -> Unit) {
     include(name)
     project(":$name").apply(block)
 }
-include("Yatoclip")
